feat: put both file-reader and file-summarizer in the same dir
