+++ knowledge/file-loader.gpt
@@ -1,0 +1,10 @@
+Name: File Loader
+Description: This tool reads any supported type of text-heavy documents(PDF, DOCX, PPTX, HTML) in the workspace, converts its content to markdown format. The parsed content will be saved to a workspace file.
+Params: Input: The workspace file to be read. This file can be of any supported text-heavy documents. Supported extensions: .pdf, .docx, .pptx, .html.
+Params: Output: File name to save the parsed content. The file name must have a ".md" extension.
+Params: know_load_metadata: Comma-delimited key=value pairs to be added to the metadata of the loaded document.
+Credential: github.com/gptscript-ai/credentials/model-provider
+Credential: ../existing-credential as knowledge
+Metadata: noUserAuth: sys.model.provider.credential,knowledge
+
+#!${GPTSCRIPT_TOOL_DIR}/bin/gptscript-go-tool load --format=markdown --flows-file=blueprint:obot --flow=obotload "ws://files/${INPUT}" "ws://files/${OUTPUT}"
